#1 implement controller and service layers for app user management

+++ back-end/src/main/java/org/hachko/poc/service/impl/AppUserManagement.java
@@ -3,6 +3,9 @@ package org.hachko.poc.service.impl;
 import java.util.List;
 
 import org.hachko.poc.dto.AppUserDto;
+import org.hachko.poc.exception.user.AppUserDuplicateEmailException;
+import org.hachko.poc.exception.user.AppUserDuplicateUserNameException;
+import org.hachko.poc.exception.user.AppUserNotFoundException;
 import org.hachko.poc.mapper.UserMapper;
 import org.hachko.poc.repository.UserRepository;
 import org.hachko.poc.service.AppUserService;
@@ -25,22 +28,42 @@ public class AppUserManagement implements AppUserService {
     }
 
     @Override
-    public AppUserDto getUserById(Long id) {
-        return userMapper.toDto(userRepository.findById(Long.valueOf(id)).orElse(null));
+    public AppUserDto getUserById(Long id) throws AppUserNotFoundException {
+        return userMapper.toDto(userRepository.findById(id).orElseThrow(
+            () -> new AppUserNotFoundException("Get : user with ID " + id + " not found."))
+        );
     }
 
     @Override
-    public AppUserDto createUser(AppUserDto userDto) {
+    public AppUserDto createUser(AppUserDto userDto) throws AppUserDuplicateUserNameException, AppUserDuplicateEmailException {
+        if(userRepository.findByUsername(userDto.getUsername()).isPresent()) {
+            throw new AppUserDuplicateUserNameException("Create : username '" + userDto.getUsername() + "' is already taken.");
+        }
+        if(userRepository.findByEmail(userDto.getEmail()).isPresent()) {
+            throw new AppUserDuplicateEmailException("Create : email '" + userDto.getEmail() + "' is already taken.");
+        }
         return userMapper.toDto(userRepository.save(userMapper.toEntity(userDto)));
     }
 
     @Override
-    public AppUserDto updateUser(AppUserDto userDto) {
+    public AppUserDto updateUser(AppUserDto userDto) throws AppUserDuplicateUserNameException, AppUserDuplicateEmailException {        
+        if(userRepository.findById(userDto.getId()).isEmpty()) {
+            throw new AppUserNotFoundException("Update : user with ID " + userDto.getId() + " not found.");
+        }
+        if(userRepository.findByUsername(userDto.getUsername()).isPresent()) {
+            throw new AppUserDuplicateUserNameException("Update : username '" + userDto.getUsername() + "' is already taken.");
+        }
+        if(userRepository.findByEmail(userDto.getEmail()).isPresent()) {
+            throw new AppUserDuplicateEmailException("Update : email '" + userDto.getEmail() + "' is already taken.");
+        }
         return userMapper.toDto(userRepository.save(userMapper.toEntity(userDto)));
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id) throws AppUserNotFoundException {
+        if(userRepository.findById(id).isEmpty()) {
+            throw new AppUserNotFoundException("Delete : user with ID " + id + " not found.");
+        }
         userRepository.deleteById(id);
     }
 }
